academic_summarizer/requirements.txt
# Core Django
django>=4.2
djangorestframework>=3.14

# Environment variables
python-dotenv>=1.0.0

# HTTP client (ScaleDown API)
requests>=2.31.0

# RAG / LLM tooling
langchain>=0.1.0
langchain-community>=0.0.1
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4

# Embeddings
sentence-transformers>=2.2.2
torch>=2.0.0

# PDF processing
pymupdf>=1.23.0
